Create component sortRepos
